+++ packages/web/runtime/index.js
@@ -11,6 +11,7 @@ import "./code-block.js";
 export * from "./dom.js";
 export * from "./events.js";
 export * from "./mount.js";
+export * from "./hydrate.js";
 export * from "./lifecycle.js";
 export * from "./router.js";
 export * from "./context.js";